Rearrange project structure
The sass load_path options do not seem to allow you to go up from the
base of the source. So, keeping the primary sartorial sass up a
directory from the source proved difficult.

However, if the scss source is on the same level, it does not need to be
prefixed with an underscore. It then still stands out as separate from
the _sass directory which is used for sass specific to the docs site.

This works for me.