Remove header/description logic from structured html tag
Sartorial includes a custom jekyll tag for taking a nested set of data
and turning it into a series of <ul> and <li> tags. Previously, this tag
includes the <h3> tag and <p> description.

This meant when looking at the index page it was sort of difficult to
see the structure of the page. This removes that logic from the custom
tag and puts it into the html.

Ideally, there would be some sort of page layout, or maybe a section tag
or something that more clearly shows the structure. But for now, between
the h3 tag and prettier, it's not that much typing/formatting to do.