Fixing horizontal scroll, in an ugly way
Apparently it's a well know html/css bug where you have horizontal
scroll but need a right margin or right padding between the area being
scrolled and the parent container.

So...for now I am just going to add 3 spaces to the longest line in the
snippet for now until this is solved.