Add NavbarStandard
I am attempting to have a few different flavors of navbar which
alternate between ease of use and most configurable.

This adds the idea of a "NavbarStandard" which has some elements
(probably a logo?) and then n number of elements that are visible on a
tablet or phone, a toggle which is used to be a hamburger menu, and then
all other elements which will make up a dropdown menu bar.