Move toggle logic into nav base, refactor navbar height
In hopes of having several different navbar modules with shared
behavior, I moved the toggle behavior into the nav based class. This may
move later as well.

More importantly, this did a good amount of refactoring in terms of how
navbar height is calculated.

Previously, you would set the navbar height and any image (presumably a
logo) would have the height set to the navbar height. Well, if the image
had padding, that would increase the height of the navbar because the
div surrounding it would increase.

The $navbar-height should really represent the actual height of the
navbar. We use this and do calculations with the spacing to do the
appropriate sizing for an image. Really, we should maybe have a class
for targeting it instead of a blanket `img` selector, but for now this
is an improvement. Another to-do would be to calculate both top/bottom
padding and left/right padding as separate variables that can be set.

This also sets the padding of the first item to 20px to match the main
content container. It's possible that is redundant and silly and we
should just remove both, but it's consistent for now.