Remove left/center/right, rename standard to split
This adds a little bit of scss to allow for a centered navbar. It also
elminates the idea of a left/centered/right navbar as I have not seen
many examples this in the real world. If it is desired we can add it
later, the only issue is that it makes collapsing somewhat difficult,
although maybe it is accomplishable with changing the ordering of items.

(How do you get all items after a checkbox, but put the logo on the
left, I think ordering would make that work?)